Move the MiSTer.ini loading logic to a separate crate and cleanup deps
We likely wont need it anymore soon, but it's useful to keep around
as there might still be a use for it (e.g. migration on first time
start).

This is a large refactor but has zero impact on actual running features.

The other major part of this is a slight cleanup of all dependencies
and cargo.